Got repo cloning to work.
